[BE] chore: manage unused file to move unused package
- each package has unused package if there are unused classes
